feat: support preview panel for recent pages
