+++ frontend/appflowy_flutter/lib/mobile/presentation/search/mobile_search_textfield.dart
@@ -150,10 +150,10 @@ class _MobileSearchTextfieldState extends State<MobileSearchTextfield> {
               child: Padding(
                 padding: const EdgeInsets.fromLTRB(4, 10, 8, 10),
                 child: FlowySvg(
-                    FlowySvgs.search_clear_m,
-                    color: theme.iconColorScheme.secondary,
-                    size: const Size.square(20),
-                  ),
+                  FlowySvgs.search_clear_m,
+                  color: theme.iconColorScheme.secondary,
+                  size: const Size.square(20),
+                ),
               ),
             )
           : null,
