+++ frontend/appflowy_flutter/lib/workspace/presentation/command_palette/widgets/search_recent_view_cell.dart
@@ -2,14 +2,18 @@ import 'package:appflowy/mobile/presentation/search/mobile_view_ancestors.dart';
 import 'package:appflowy/startup/startup.dart';
 import 'package:appflowy/workspace/application/action_navigation/action_navigation_bloc.dart';
 import 'package:appflowy/workspace/application/action_navigation/navigation_action.dart';
+import 'package:appflowy/workspace/application/recent/recent_views_bloc.dart';
 import 'package:appflowy/workspace/application/view/view_ext.dart';
 import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:appflowy_ui/appflowy_ui.dart';
-import 'package:flowy_infra_ui/widget/spacing.dart';
+import 'package:flowy_infra/theme_extension.dart';
+import 'package:flowy_infra_ui/flowy_infra_ui.dart';
+import 'package:flowy_infra_ui/style_widget/hover.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class SearchRecentViewCell extends StatelessWidget {
+class SearchRecentViewCell extends StatefulWidget {
   const SearchRecentViewCell({
     super.key,
     required this.icon,
@@ -22,50 +26,82 @@ class SearchRecentViewCell extends StatelessWidget {
   final VoidCallback onSelected;
 
   @override
+  State<SearchRecentViewCell> createState() => _SearchRecentViewCellState();
+}
+
+class _SearchRecentViewCellState extends State<SearchRecentViewCell> {
+  final focusNode = FocusNode();
+
+  ViewPB get view => widget.view;
+
+  @override
+  void dispose() {
+    focusNode.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     final theme = AppFlowyTheme.of(context),
         textColor = theme.textColorScheme.primary;
     final sapceM = theme.spacing.m, spaceL = theme.spacing.l;
+    final bloc = context.read<RecentViewsBloc>(), state = bloc.state;
+    final hoveredView = state.hoveredView;
+    final hovering = hoveredView == view;
 
-    return AFBaseButton(
-      borderRadius: sapceM,
-      padding: EdgeInsets.symmetric(vertical: spaceL, horizontal: sapceM),
-      backgroundColor: (context, isHovering, disable) {
-        if (isHovering) {
-          return Theme.of(context).colorScheme.secondary;
-        }
-        return theme.fillColorScheme.transparent;
-      },
-      borderColor: (context, isHovering, disable, isFocused) =>
-          theme.fillColorScheme.transparent,
-      builder: (ctx, isHovering, disable) => Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          icon,
-          HSpace(12),
-          Flexible(
-            child: Column(
+    return GestureDetector(
+      behavior: HitTestBehavior.opaque,
+      onTap: () => _handleSelection(view.id),
+      child: Focus(
+        focusNode: focusNode,
+        onKeyEvent: (node, event) {
+          if (event is! KeyDownEvent) return KeyEventResult.ignored;
+          if (event.logicalKey == LogicalKeyboardKey.enter) {
+            _handleSelection(view.id);
+            return KeyEventResult.handled;
+          }
+          return KeyEventResult.ignored;
+        },
+        onFocusChange: (hasFocus) {
+          if (hasFocus && !hovering) {
+            bloc.add(RecentViewsEvent.hoverView(view));
+          }
+        },
+        child: FlowyHover(
+          onHover: (value) {
+            if (hoveredView == view) return;
+            bloc.add(RecentViewsEvent.hoverView(view));
+          },
+          style: HoverStyle(
+            borderRadius: BorderRadius.circular(8),
+            hoverColor: Theme.of(context).colorScheme.secondary,
+            foregroundColorOnHover: AFThemeExtension.of(context).textColor,
+          ),
+          isSelected: () => hovering,
+          child: Padding(
+            padding: EdgeInsets.symmetric(vertical: spaceL, horizontal: sapceM),
+            child: Row(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Text(
-                  view.nameOrDefault,
-                  style: theme.textStyle.body.standard(color: textColor),
+                widget.icon,
+                HSpace(12),
+                Flexible(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        view.nameOrDefault,
+                        style: theme.textStyle.body.standard(color: textColor),
+                      ),
+                      buildPath(theme),
+                    ],
+                  ),
                 ),
-                buildPath(theme),
               ],
             ),
           ),
-        ],
+        ),
       ),
-      onTap: () {
-        onSelected();
-
-        getIt<ActionNavigationBloc>().add(
-          ActionNavigationEvent.performAction(
-            action: NavigationAction(objectId: view.id),
-          ),
-        );
-      },
     );
   }
 
@@ -74,6 +110,16 @@ class SearchRecentViewCell extends StatelessWidget {
       create: (context) => ViewAncestorBloc(view.id),
       child: BlocBuilder<ViewAncestorBloc, ViewAncestorState>(
         builder: (context, state) => state.buildPath(context),
+      ),
+    );
+  }
+
+  /// Helper to handle the selection action.
+  void _handleSelection(String id) {
+    widget.onSelected();
+    getIt<ActionNavigationBloc>().add(
+      ActionNavigationEvent.performAction(
+        action: NavigationAction(objectId: id),
       ),
     );
   }
